Add empowered on damage absorbed

+++ Unity/Assets/Script/Gameplay/Entities/Modifier/Behaviours/IModifierStack.cs
@@ -2,6 +2,6 @@
 {
     public interface IModifierStack
     {
-        public int CurrentStack { get; set; }
+        public float CurrentStack { get; set; }
     }
 }
